Avoid ambiguous call of pow()
git-svn-id: https://www.ibr.cs.tu-bs.de/svn/emma/trunk/src/ibrdtn@33884 1cffb75b-d306-6187-5d55-5a277715007a

+++ daemon/src/routing/RetransmissionExtension.cpp
@@ -165,7 +165,7 @@ namespace dtn
 		{
 			_count++;
 			_timestamp = dtn::utils::Clock::getTime();
-			_timestamp += pow(retry, _count -1);
+			_timestamp += pow((float)retry, (float)_count -1);
 
 			return (*this);
 		}
@@ -174,7 +174,7 @@ namespace dtn
 		{
 			_count++;
 			_timestamp = dtn::utils::Clock::getTime();
-			_timestamp += pow(retry, _count -1);
+			_timestamp += pow((float)retry, (float)_count -1);
 
 			return (*this);
 		}
